requirements.txt
flask
pytesseract
pillow
opencv-python
gunicorn
pytest
